Improve naming in various places.
- Split catalog module into surgery and checkup.
- Use unhealthy instead of abberation
- Use healthcheck instead of checkup
- shorten zopectl command name
