+++ ftw/catalogdoctor/command.py
@@ -1,6 +1,6 @@
 from __future__ import print_function
-from ftw.catalogdoctor.catalog import CatalogCheckup
-from ftw.catalogdoctor.catalog import CatalogDoctor
+from ftw.catalogdoctor.healthcheck import CatalogHealthCheck
+from ftw.catalogdoctor.surgery import CatalogDoctor
 from Products.CMFCore.utils import getToolByName
 from Products.CMFPlone.interfaces import IPloneSiteRoot
 from Testing.makerequest import makerequest
@@ -44,30 +44,30 @@ class ConsoleOutput(object):
         print(msg, file=sys.stderr)
 
 
-def checkup_command(portal_catalog):
-    result = CatalogCheckup(catalog=portal_catalog).run()
+def healthcheck_command(portal_catalog):
+    result = CatalogHealthCheck(catalog=portal_catalog).run()
     result.write_result(formatter=ConsoleOutput())
     return result
 
 
 def surgery_command(portal_catalog):
-    result = checkup_command(portal_catalog)
+    result = healthcheck_command(portal_catalog)
 
     formatter = ConsoleOutput()
     there_is_nothing_we_can_do = []
 
-    for aberration in result.get_aberrations():
-        doctor = CatalogDoctor(result.catalog, aberration)
+    for unhealthy_rid in result.get_unhealthy_rids():
+        doctor = CatalogDoctor(result.catalog, unhealthy_rid)
         if doctor.can_perform_surgery():
             result = doctor.perform_surgery()
             formatter.info(result)
         else:
-            there_is_nothing_we_can_do.append(aberration)
+            there_is_nothing_we_can_do.append(unhealthy_rid)
 
     if there_is_nothing_we_can_do:
-        formatter.info('The following aberrations could not be fixed')
-        for aberration in there_is_nothing_we_can_do:
-            aberration.write_result(formatter)
+        formatter.info('The following unhealthy rids could not be fixed')
+        for unhealthy_rid in there_is_nothing_we_can_do:
+            unhealthy_rid.write_result(formatter)
 
 
 def doctor_cmd(app, args):
@@ -87,14 +87,15 @@ def doctor_cmd(app, args):
         help='Dryrun, do not commit changes')
 
     commands = parser.add_subparsers()
-    checkup = commands.add_parser(
-        'checkup',
-        help='Run a checkup for portal_catalog.')
-    checkup.set_defaults(func=checkup_command)
+    healthcheck = commands.add_parser(
+        'healthcheck',
+        help='Run a health check for portal_catalog.')
+    healthcheck.set_defaults(func=healthcheck_command)
 
     surgery = commands.add_parser(
         'surgery',
-        help='Run a checkup and perform surgery for aberrations in portal_catalog.')
+        help='Run a healthcheck and perform surgery for unhealthy rids in '
+             'portal_catalog.')
     surgery.set_defaults(func=surgery_command)
 
     args = parser.parse_args(args)
